requirements.txt
Django==4.2.10
dj-database-url==2.1.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
